Add home row modifiers; redo lower layer
This is still very similar to what I did on Planck. This could be a bad
idea, given that in the Planck I did a few compromises, and the
placement of the punctuation is weird. Also, the numbers are in a very
uncommon way due wanting to have a "one handed mode" for numbers, as
that's something I've wanted for some games. In the Planck the numbers
have to be put somewhere in a layer, but for a Sofle the numbers have
their own keys.