Big revamp of Adjust, to prepare the lower ones
Replanned the design of the layers.

Function keys on the Raise layer (third) are problematic because in the
Plank I had to make a few of them double duty with tap-dance, with
F-keys sharing mouse scroll ones.

Moving functions to the Adjust ones solves that, with the caveat that in
the Sofle I can actually use the number row for the function keys in the
lower layer. So I can still do single handed activation of F2, F3, etc.
Maybe even Alt+F4!

With the extra row on adjust, I have added Windows+number in the number
row, so I can do window switching in Plasma a bit more pleasant, and
maybe on a Mac as well. The Raise layer has received Ctrl+Fkey,
something that I should do more as well in Plasma.

The adjust layer has received all the multimedia keys, with volume
up/down, and next/previous song in nice HJKL positions, inspired by
Miryoku.

Also inspired by Miryoku is the use of "layer lock/pin" and "layer
reset" keys. It's just a pair of `TO(x)` mappings that set the current
layer as base layer (so it locks/pins the layer when releasing the layer
keys), with an additional key to go back to the base layer, which is
like "resetting" the lock. The one to reset should be where Escape is,
so is natural to hit that frantically and repeatedly without fear.