Almost final part of the refactor
Removed the cursor keys from the LOWER layer, and rearranged the symbols
taking advantage of new ideas. The <-+> keys resemble the movements of
typical HJKL, so they are moved there, which leaves the backslash where
the regular slash would be in the main layer. Comma and period are now
in the same position in the lower layer and the base one, which is
useful to accommodate the use of typing IPs, sequences of numbers, etc.

The RAISE layer is now largely inspired by the navigation layer of
Miryoku, with shortcuts to ctrl+cursor, as an experiment, and navigation
keys on top instead of below. This layer also has gained a few
multimedia keys, as it's easy enough to press with one hand the RAISE
button and the keys to the side.