moCkiNG casE aka sPONgEbOB Case IMplEMENTaTiOn
Implement a "mocking case" mode, AKA SpongeBob Case. It tries to be
exclusive with the CAPS_WORD feature, but it's still colliding a bit
with CapsLock. The OLED screen message seems to work fine. All the tree
modes could cooperate a bit more, but due to how differently are
implemented, it's probably not worth the effort.