Better use of the lowermost right keys
I can't have a full arrow set on a Sofle, like on the Planck, given the
lack of just one key. But I still miss out on having at least left and
right for scrolling through videos when my hands are not on top of the
keyboard.

Since I'm also starting to need an AltGr key more and more for typing in
non-English languages, let's try having it there.